+++ config/services.php
@@ -230,6 +230,7 @@ return static function (ContainerConfigurator $container, ContainerBuilder $buil
                 ->args([
                     service(CommandRunner::class),
                     $symfonyConfig['flex']['command'],
+                    $symfonyConfig['flex']['timeout'],
                 ])
                 ->tag('jmonitor.symfony.component_collector', ['index' => 'flex_recipes'])
             ;